feat: 容器shell对接webconsole
