+++ containers/Compute/views/pod-container/components/List.vue
@@ -101,6 +101,24 @@ export default {
     this.list.fetchData()
   },
   methods: {
+    openWebConsole (data) {
+      let href = ''
+      if (process.env.NODE_ENV === 'development') {
+        href = `${this.$appConfig.webConsolePath}${data.access_url.replace(/^.*?web-console\//, '')}`
+      } else {
+        href = data.access_url
+      }
+      window.open(`${href}&session_id=${data.session}`)
+    },
+    async fetchConnectUrl (obj) {
+      const { data } = await new this.$Manager('webconsole', 'v1').objectRpc({
+        methodname: 'DoContainerExec',
+        params: {
+          container_id: obj.id,
+        },
+      })
+      return Promise.resolve(data)
+    },
     handleOpenSidepage (row, tab) {
       this.sidePageTriggerHandle(this, 'VmPodContainerSidePage', {
         id: row.id,
